add utilities and add some refactors
